fix: split list pages into layout + index routes for nested routing
